Proposed changes to API
1. made load_data, saveoutput and _normalize_desc method the concrete ones, because they shouldn't depend on the cluster or diversity method of selection
2. add __init__ method to a SelectionBase abstract class, because there are properties that can be the same for both Cluster and Dissimilarity selectors
3. rewrite DessimilaritySelector class based on the changes of the parent class (even thought it's an abstract class)
